refactor(ocale): migrate game-passes wire code to by-domain layout
Move builders, parsers, wire types, rate-limit constants, and parsed
shapes for the `/game-passes/v1/...` Operations from
`src/resources/game-passes/` to a new per-shape sub-tree under
`src/domains/game-passes/game-passes/`, per the by-Domain layout in
ADR-022. The Resource folder becomes a thin facade: `client.ts` imports
from the Domain sub-tree and `index.ts` re-exports the public types
from there too.

Wire requests, parsed shapes, and rate-limit operation keys
(`game-passes.get`, `game-passes.create`) are preserved end-to-end.
