+++ packages/open-cloud/src/domains/game-passes/game-passes/parsers.ts
@@ -1,8 +1,11 @@
-import type { HttpResponse } from "../../client/types.ts";
-import { ApiError } from "../../errors/api-error.ts";
-import { copyPriceInformation, isPriceInformationLike } from "../../internal/price-information.ts";
-import { isRecord } from "../../internal/utils/is-record.ts";
-import type { Result } from "../../types.ts";
+import type { HttpResponse } from "../../../client/types.ts";
+import { ApiError } from "../../../errors/api-error.ts";
+import {
+	copyPriceInformation,
+	isPriceInformationLike,
+} from "../../../internal/price-information.ts";
+import { isRecord } from "../../../internal/utils/is-record.ts";
+import type { Result } from "../../../types.ts";
 import type { GamePass } from "./types.ts";
 import type { GamePassConfigV2, PricingFeatureWire } from "./wire.ts";
