+++ packages/open-cloud/src/resources/game-passes/index.ts
@@ -1,8 +1,8 @@
-export { GamePassesClient } from "./client.ts";
 export type {
 	CreateGamePassParameters,
 	GamePass,
 	GamePassPrice,
 	GamePassPricingFeature,
 	GetGamePassParameters,
-} from "./types.ts";
+} from "../../domains/game-passes/game-passes/types.ts";
+export { GamePassesClient } from "./client.ts";
