+++ packages/open-cloud/tests/conformance/game-passes.spec.ts
@@ -1,4 +1,4 @@
-import { parseGamePassResponse } from "#src/resources/game-passes/parsers";
+import { parseGamePassResponse } from "#src/domains/game-passes/game-passes/parsers";
 import { assert, describe, expect, it } from "vitest";
 
 import {
